Adding KeySym enum generator
This will generate the KeySym enum from the c header file

+++ src/integrationTest/java/com/github/moaxcp/x11client/tinywm/TinyWindowManager.java
@@ -18,7 +18,7 @@ public class TinyWindowManager {
   public void start() throws IOException {
     try(X11Client client = X11Client.connect(new DisplayName(display))) {
       client.send(GrabKey.builder()
-        .key((byte) client.keySymToKeyCode(KeySym.getByName("F1").get().getValue()))
+        .key(client.keySymToKeyCode(KeySym.getByName("F1").get().getValue()))
         .modifiersEnable(ModMask.ONE)
         .grabWindow(client.getRoot(0))
         .ownerEvents(true)
